Refactor ingredients data structure
