+++ lib/components/material_item.dart
@@ -7,8 +7,8 @@ import "package:material_symbols_icons/symbols.dart";
 import "../db/bookmark_db_extension.dart";
 import "../i18n/strings.g.dart";
 import "../models/bookmark.dart";
-import "../models/character_ingredients.dart";
 import "../models/common.dart";
+import "../models/ingredients.dart";
 import "../models/material.dart";
 import "../models/material_bookmark_frame.dart";
 import "../providers/database_provider.dart";
